Remove search module and update home module
