+++ lib/modules/app/base_app_module.dart
@@ -1,6 +1,5 @@
 import 'package:flutter_base_template_1/modules/detail/detail_module.dart';
 import 'package:flutter_base_template_1/modules/home/home_module.dart';
-import 'package:flutter_base_template_1/modules/search/search_module.dart';
 import 'package:flutter_base_template_1/networking/constants/network_constants.dart';
 import 'package:flutter_base_template_1/networking/http_client.dart';
 import 'package:flutter_base_template_1/networking/models/app_dio.dart';
@@ -28,15 +27,10 @@ class BaseAppModule extends Module {
           BaseAppModuleRoutes.detailPage,
           module: DetailModule(),
         ),
-        ModuleRoute(
-          BaseAppModuleRoutes.searchPage,
-          module: SearchModule(),
-        ),
       ];
 }
 
 class BaseAppModuleRoutes {
   static const String homePage = '/';
   static const String detailPage = '/detail/';
-  static const String searchPage = '/search/';
 }
