+++ lib/modules/home/repositories/home_repo.dart
@@ -1,10 +1,16 @@
-import 'package:flutter_base_template_1/modules/home/models/home_events_response.dart';
+import 'package:flutter_base_template_1/modules/home/models/events_response.dart';
 import 'package:flutter_base_template_1/networking/retrofit/home_api_client.dart';
 
 class HomeRepo {
-  final HomeApiClient _homeApiClient = HomeApiClient.withAuthenticatedDio();
+  final HomeApiClient _homeApiClient = HomeApiClient.withDefaultDio();
 
-  Future<HomeEventsResponse> getEventByEventId({required int eventId}) async {
-    return _homeApiClient.getEventByEventId(eventId);
+  Future<EventsResponse> fetchEvents({
+    required String searchString,
+    required int page,
+  }) async {
+    return _homeApiClient.getEventsData(
+      search: searchString,
+      page: page,
+    );
   }
 }
